Database: Case counting system change
The old system, under extreme race conditions can bug out and produce 2 cases with the same case id (not global case id).

Here, we change it to instead count the cases, as for case removal we have a removed field.

A unique constraint is also implemented in ModerationCase as a extra-measure